refactor: move identity index/template descriptors to non component packages
