+++ search/search-client-query-transformer/src/main/java/io/camunda/search/clients/transformers/filter/UserFilterTransformer.java
@@ -9,10 +9,10 @@ package io.camunda.search.clients.transformers.filter;
 
 import static io.camunda.search.clients.query.SearchQueryBuilders.and;
 import static io.camunda.search.clients.query.SearchQueryBuilders.term;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.EMAIL;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.KEY;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.NAME;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.USERNAME;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.EMAIL;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.KEY;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.NAME;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.USERNAME;
 
 import io.camunda.search.clients.query.SearchQuery;
 import io.camunda.search.filter.UserFilter;
