+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/GroupIndex.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.usermanagement.index;
+package io.camunda.webapps.schema.descriptors.index;
 
 import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
 import io.camunda.webapps.schema.descriptors.ComponentNames;
@@ -14,31 +14,21 @@ import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.Enti
 import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.IdentityJoinRelationshipType;
 import java.util.Optional;
 
-public class RoleIndex extends AbstractIndexDescriptor implements Prio5Backup {
-  public static final String INDEX_NAME = "role";
-  public static final String INDEX_VERSION = "8.7.0";
+public class GroupIndex extends AbstractIndexDescriptor implements Prio5Backup {
 
+  public static final String INDEX_NAME = "group";
+  public static final String INDEX_VERSION = "8.7.0";
   public static final String KEY = "key";
-  public static final String NAME = "name";
   public static final String MEMBER_KEY = "memberKey";
+  public static final String NAME = "name";
   public static final String JOIN = "join";
 
   public static final EntityJoinRelationFactory JOIN_RELATION_FACTORY =
       new EntityJoinRelationFactory(
-          IdentityJoinRelationshipType.ROLE, IdentityJoinRelationshipType.MEMBER);
+          IdentityJoinRelationshipType.GROUP, IdentityJoinRelationshipType.MEMBER);
 
-  public RoleIndex(final String indexPrefix, final boolean isElasticsearch) {
+  public GroupIndex(final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
-  }
-
-  @Override
-  public String getVersion() {
-    return INDEX_VERSION;
-  }
-
-  @Override
-  public String getComponentName() {
-    return ComponentNames.CAMUNDA.toString();
   }
 
   @Override
@@ -49,5 +39,15 @@ public class RoleIndex extends AbstractIndexDescriptor implements Prio5Backup {
   @Override
   public Optional<String> getTenantIdField() {
     return Optional.empty();
+  }
+
+  @Override
+  public String getVersion() {
+    return INDEX_VERSION;
+  }
+
+  @Override
+  public String getComponentName() {
+    return ComponentNames.CAMUNDA.toString();
   }
 }
