+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/MappingIndex.java
@@ -5,32 +5,24 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.usermanagement.index;
+package io.camunda.webapps.schema.descriptors.index;
 
 import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
 import io.camunda.webapps.schema.descriptors.ComponentNames;
 import io.camunda.webapps.schema.descriptors.backup.Prio5Backup;
 import java.util.Optional;
 
-public class PersistentWebSessionIndexDescriptor extends AbstractIndexDescriptor
-    implements Prio5Backup {
-
-  public static final String INDEX_NAME = "web-session";
+public class MappingIndex extends AbstractIndexDescriptor implements Prio5Backup {
+  public static final String INDEX_NAME = "mapping";
   public static final String INDEX_VERSION = "8.7.0";
 
-  public PersistentWebSessionIndexDescriptor(
-      final String indexPrefix, final boolean isElasticsearch) {
+  public static final String MAPPING_KEY = "mappingKey";
+  public static final String CLAIM_NAME = "claimName";
+  public static final String CLAIM_VALUE = "claimValue";
+  public static final String NAME = "name";
+
+  public MappingIndex(final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
-  }
-
-  @Override
-  public String getIndexName() {
-    return INDEX_NAME;
-  }
-
-  @Override
-  public Optional<String> getTenantIdField() {
-    return Optional.empty();
   }
 
   @Override
@@ -41,5 +33,15 @@ public class PersistentWebSessionIndexDescriptor extends AbstractIndexDescriptor
   @Override
   public String getComponentName() {
     return ComponentNames.CAMUNDA.toString();
+  }
+
+  @Override
+  public String getIndexName() {
+    return INDEX_NAME;
+  }
+
+  @Override
+  public Optional<String> getTenantIdField() {
+    return Optional.empty();
   }
 }
