+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/PersistentWebSessionIndexDescriptor.java
@@ -5,29 +5,21 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.usermanagement.index;
+package io.camunda.webapps.schema.descriptors.index;
 
 import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
 import io.camunda.webapps.schema.descriptors.ComponentNames;
 import io.camunda.webapps.schema.descriptors.backup.Prio5Backup;
-import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.EntityJoinRelationFactory;
-import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.IdentityJoinRelationshipType;
 import java.util.Optional;
 
-public class GroupIndex extends AbstractIndexDescriptor implements Prio5Backup {
+public class PersistentWebSessionIndexDescriptor extends AbstractIndexDescriptor
+    implements Prio5Backup {
 
-  public static final String INDEX_NAME = "group";
+  public static final String INDEX_NAME = "web-session";
   public static final String INDEX_VERSION = "8.7.0";
-  public static final String KEY = "key";
-  public static final String MEMBER_KEY = "memberKey";
-  public static final String NAME = "name";
-  public static final String JOIN = "join";
 
-  public static final EntityJoinRelationFactory JOIN_RELATION_FACTORY =
-      new EntityJoinRelationFactory(
-          IdentityJoinRelationshipType.GROUP, IdentityJoinRelationshipType.MEMBER);
-
-  public GroupIndex(final String indexPrefix, final boolean isElasticsearch) {
+  public PersistentWebSessionIndexDescriptor(
+      final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
   }
