+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/RoleIndex.java
@@ -5,23 +5,29 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.usermanagement.index;
+package io.camunda.webapps.schema.descriptors.index;
 
 import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
 import io.camunda.webapps.schema.descriptors.ComponentNames;
 import io.camunda.webapps.schema.descriptors.backup.Prio5Backup;
+import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.EntityJoinRelationFactory;
+import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.IdentityJoinRelationshipType;
 import java.util.Optional;
 
-public class MappingIndex extends AbstractIndexDescriptor implements Prio5Backup {
-  public static final String INDEX_NAME = "mapping";
+public class RoleIndex extends AbstractIndexDescriptor implements Prio5Backup {
+  public static final String INDEX_NAME = "role";
   public static final String INDEX_VERSION = "8.7.0";
 
-  public static final String MAPPING_KEY = "mappingKey";
-  public static final String CLAIM_NAME = "claimName";
-  public static final String CLAIM_VALUE = "claimValue";
+  public static final String KEY = "key";
   public static final String NAME = "name";
+  public static final String MEMBER_KEY = "memberKey";
+  public static final String JOIN = "join";
 
-  public MappingIndex(final String indexPrefix, final boolean isElasticsearch) {
+  public static final EntityJoinRelationFactory JOIN_RELATION_FACTORY =
+      new EntityJoinRelationFactory(
+          IdentityJoinRelationshipType.ROLE, IdentityJoinRelationshipType.MEMBER);
+
+  public RoleIndex(final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
   }
